make: Autospec creation for version 4.2.1
Adding the fix from upstream for http://savannah.gnu.org/bugs/?func=detailitem&item_id=50902

The basic issue is that the perl test driver says
     require "test_driver.pl";

but the magic @INC array that is built into perl doesn't have "." as a
place to search. I would have fixed the issue with just pushing "."
onto @INC rather than adding a dependency on FindBin, but as this is a
known issue and has a fix accepted upstream it seems better to accept
it.

Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>